Consolidate response writing
