+++ GPTifier/CMakeLists.txt
@@ -59,7 +59,6 @@ add_executable(
   "src/help_messages.cpp"
   "src/main.cpp"
   "src/params.cpp"
-  "src/responses.cpp"
   "src/utils.cpp"
 )
 target_link_libraries(
